DHIS2-2041: avoid jQuery parsererror on PATCH request
If the server replies with 200 Success but empty response,
jQuery throws a parsererror error, as it's trying to parse an
empty JSON response.
Since this PATCH request is never expected to have actual
content in the response, force the dataType to 'text' to
avoid the problem.
Ideally (as it currently happens in 2.27/demo) the server
should respond with a 204 No content, which apparently
prevents jQuery from trying to parse an empty response.

Opened a bug for the server, but added this fix anyway.